[Chore] Relocate directory
