+++ todolist/src/components/column/card/CardList.jsx
@@ -1,0 +1,14 @@
+import CardWrap from "./CardWrap";
+
+function CardList(props) {
+  const { cards, onDelete, onUpdate } = props;
+  return (
+    <ul className="card-wrap">
+      {cards.map((card) => (
+        <CardWrap key={card.id} card={card} onDelete={onDelete} onUpdate={onUpdate} />
+      ))}
+    </ul>
+  );
+}
+
+export default CardList;
